+++ variationSteps.cu
@@ -21,7 +21,7 @@
     }
 }*/
 
-void steps2Variation(int* variation, int* steps, Kernel* kernel)
+void steps2Variation(uint64_t* variation, int* steps, Kernel* kernel)
 {
     *variation = 0;
     int attrStride = 1;
